refactor(domain): split types into enum files and update imports
create separate enum files for TransactionType and RepeatFrequency, delete the combined value-objects file, and update imports in category and transaction entities
