+++ src/domain/enum/repeat-frequency.ts
@@ -1,3 +1,1 @@
-export type TransactionType = 'EXPENSE' | 'INCOME';
-
 export type RepeatFrequency = 'MONTHS' | 'WEEKS' | 'DAYS';
