+++ src/domain/enum/transaction-type.ts
@@ -1,0 +1,1 @@
+export type TransactionType = 'EXPENSE' | 'INCOME';
